IIT Madras Authorization Letter For Data Usage
The IIT Madras Authorization Letter for Data Usage is an official document that supports the student's request to collect and analyze data from industry or business establishments as part of their Capstone project in the Business Data Management course. It is ensured that the data collected will be used strictly for academic purposes and will not be shared with others.